chore: move react-queries into separate files
